feat: Add support for user image messages in chat
This commit adds support for user image messages in the chat feature. It includes modifications to the `ChatModel` enum to include a new message type `USER_IMAGE_MESSAGE`. Additionally, the `ChatMessageComponent` has been updated to display user image messages with the appropriate styling. The `AppComponent` template has also been modified to include an input field for uploading images and displaying the uploaded image.